homework2

+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/Application.java
@@ -12,14 +12,14 @@ public class Application {
         } else {
             System.out.println("Metoda sum nie działa poprawnie dla liczb " + a + " i " + b);
         }
-        int substractResult = calculator.substract(a, b);
+        double substractResult = calculator.substract(a, b);
         boolean correct1 = ResultChecker.assertEquals(-3, substractResult);
         if (correct1) {
             System.out.println("Metoda substract działą poprawnie dla liczb " + a + " i " + b);
         } else {
             System.out.println("Metoda substract nie działą poprawnie dla " + a + " i " + b);
         }
-        int squareResult = calculator.square(a);
+        double squareResult = calculator.square(a);
         boolean correct2 = ResultChecker.assertEquals(25,squareResult);
         if (correct2) {
             System.out.println("Metoda square działa poprawnie dla liczby " + a);
